backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.77.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
google-generativeai==0.8.3
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
requests>=2.31.0

# Additional dependencies for new features
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
